requirements.txt
aiohttp
aiosignal
async-timeout
attrs
beautifulsoup4
Brotli
certifi
cffi
charset-normalizer
discord.py
frozenlist
idna
lyricsgenius
multidict
mutagen
pycparser
pycryptodomex
PyNaCl
requests
soupsieve
urllib3
websockets
yarl
yt-dlp
python-dotenv

isort
black
